+++ src/components/CodeHighlight/index.ts
@@ -1,3 +1,3 @@
-// CodeHighlight from copy-ui @ 2025-10-05.
+// CodeHighlight from copy-ui @ 2025-10-13.
 
 export { CodeHighlight } from "./CodeHighlight";