+++ src/components/CodeHighlight/CHANGELOG.md
@@ -1,3 +1,4 @@
 - 2025-09-12: Initial version.
 - 2025-09-26: Add support for TSX language in CodeHighlight.
 - 2025-10-05: Add `withLineNumbers` prop to CodeHighlight component.
+- 2025-10-11: Update line-numbers' style.